+++ include/api/JsonAPI.h
@@ -99,6 +99,26 @@ signals:
 	///
 	void forwardJsonMessage(QJsonObject);
 
+	///
+	/// Signal emits whenever a suspend/resume request for all instances should be forwarded
+	///
+	void suspendAll(bool isSuspend);
+
+	///
+	/// Signal emits whenever a toggle suspend/resume request for all instances should be forwarded
+	///
+	void toggleSuspendAll();
+
+	///
+	/// Signal emits whenever a idle mode request for all instances should be forwarded
+	///
+	void idleAll(bool isIdle);
+
+	///
+	/// Signal emits whenever a toggle idle/working mode request for all instances should be forwarded
+	///
+	void toggleIdleAll();
+
 private:
 	// true if further callbacks are forbidden (http)
 	bool _noListener;
@@ -297,6 +317,12 @@ private:
 	/// @param message the incoming message
 	///
 	void handleServiceCommand(const QJsonObject &message, const QString &command, int tan);
+
+	/// Handle an incoming JSON message for actions related to the overall Hyperion system
+	///
+	/// @param message the incoming message
+	///
+	void handleSystemCommand(const QJsonObject &message, const QString &command, int tan);
 
 	///
 	/// Handle an incoming JSON message of unknown type