+++ src/hyperion-remote/JsonConnection.h
@@ -100,6 +100,36 @@ public:
 	QString getSysInfo();
 
 	///
+	/// Suspend Hyperion. Stop all instances and components
+	///
+	void suspend();
+
+	///
+	/// Resume Hyperion. Start all instances and components
+	///
+	void resume();
+
+	///
+	/// Toggle between Suspend and Resume
+	///
+	void toggleSuspend();
+
+	///
+	///  Put Hyperion in Idle mode, i.e. all instances, components will be disabled besides the output processing (LED-devices, smoothing).
+	///
+	void idle();
+
+	///
+	/// Toggle between Idle and Working mode
+	///
+	void toggleIdle();
+
+	///
+	/// Restart Hyperion
+	///
+	void restart();
+
+	///
 	/// Clear the given priority channel
 	///
 	/// @param priority The priority