+++ src/hyperiond/systray.h
@@ -12,6 +12,7 @@
 
 #include <hyperion/Hyperion.h>
 #include <hyperion/HyperionIManager.h>
+#include "SuspendHandler.h"
 
 class HyperionDaemon;
 
@@ -61,6 +62,8 @@ private:
 
 	QAction          *quitAction;
 	QAction          *restartAction;
+	QAction          *suspendAction;
+	QAction          *resumeAction;
 	QAction          *startAction;
 	QAction          *stopAction;
 	QAction          *colorAction;
@@ -75,9 +78,12 @@ private:
 #if defined(ENABLE_EFFECTENGINE)
 	QMenu            *_trayIconEfxMenu;
 #endif
+	QMenu            *_trayIconSystemMenu;
 	QColorDialog      _colorDlg;
 	HyperionDaemon   *_hyperiond;
 	Hyperion         *_hyperion;
 	HyperionIManager *_instanceManager;
 	quint16           _webPort;
+
+	SuspendHandler   *_suspendHandler;
 };